% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDbsSignatureProfiles.R
\docType{data}
\name{DBS_SIGNATURE_PROFILES}
\alias{DBS_SIGNATURE_PROFILES}
\title{DBS signature profiles}
\format{
An object of class \code{data.frame} with 78 rows and 11 columns.
}
\usage{
data(DBS_SIGNATURE_PROFILES)
}
\description{
A matrix
rows: dinucleotide base substitution (DBS) context
cols: DBS signatures
}
\details{
Source: https://cancer.sanger.ac.uk/sigs-assets-20/COSMIC_Mutational_Signatures_v3.1.xlsx
}
\keyword{datasets}
