% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDbsTypes.R
\docType{data}
\name{DBS_TYPES}
\alias{DBS_TYPES}
\title{Look up table for doublet base substitution types and reverse complement sequences}
\format{
An object of class \code{data.frame} with 78 rows and 3 columns.
}
\usage{
data(DBS_TYPES)
}
\description{
Excel sheet source: https://cancer.sanger.ac.uk/signatures/DBS_vignettes/DBS-doublet-base-substitution-classification.xlsx
}
\keyword{datasets}
