% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitToSignatures.R
\name{eps}
\alias{eps}
\title{Linear least-squares (non-negative) fitting}
\usage{
eps(x = 1)
}
\arguments{
\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are the mutation
contexts and the columns are  the mutational signatures.}

\item{mut.contexts}{A vector of mutation contexts for fitting}
}
\value{
A list containing, the absolute contribution of each signature (i.e. the number of mutations contributing to
each mutational signature), and the residual (amount of error from the fitting).
}
\description{
Calculate the contribution of each mutational signature in a sample, given a vector of mutation contexts.
The eps() and lsqnonneg() functions have been extracted and adapted from the pracma R package.
}
