% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsForHrdClassifier.R
\name{extractSigsForHrdClassifier}
\alias{extractSigsForHrdClassifier}
\title{Extract signatures in the format compatible with the random forest HRD classifier}
\usage{
extractSigsForHrdClassifier(vcf.snv, vcf.indel, vcf.sv,
  sample.name = "sample")
}
\arguments{
\item{vcf.snv}{Path to the vcf file containing SNVs}

\item{vcf.indel}{Path to the vcf file containing indels}

\item{vcf.sv}{Path to the vcf file containing SVs (Manta vcf)}

\item{sample.name}{The name of the sample as a character. Defaults to 'sample' if none is provided.}
}
\value{
A 1-column data frame containing the mutational signature contributions
}
\description{
Extract signatures in the format compatible with the random forest HRD classifier
}
\examples{
extractSigsForHrdClassifier(
   vcf.snv = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010/analysis/Arne/DR-10_update/scripts/SNV2//CPCT02010350R_CPCT02010350T_PASS_SNV.vcf.gz',
   vcf.indel = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010/analysis/Arne/DR-10_update/scripts/INDEL2//CPCT02010350R_CPCT02010350T_PASS_indel.vcf.gz',
   vcf.sv = '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/HMF_data/DR010-update/data//170217_HMFregCPCT_FR12244678_FR12244592_CPCT02010350/CPCT02010350R_CPCT02010350T_somaticSV_bpi.vcf.gz',
   sample.name = 'CPCT02010350R_CPCT02010350T'
)
}
