% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsIndel.R
\name{extractSigsIndel}
\alias{extractSigsIndel}
\title{Extract indel signatures}
\usage{
extractSigsIndel(vcf.file, sample.name = NULL,
  ref.genome = DEFAULT_GENOME, indel.len.cap = 5, n.bases.mh.cap = 5,
  ...)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{sample.name}{If a character is provided, the header for the output matrix will be named to this. If none is
provided, the basename of the vcf file will be used.}

\item{ref.genome}{A character naming the BSgenome reference genome. Default is "BSgenome.Hsapiens.UCSC.hg19". If another
reference genome is indicated, it will also need to be installed.}

\item{indel.len.cap}{Specifies the max indel sequence length to consider when counting 'repeat' and 'none' contexts.
Counts of longer indels will simply be binned to the counts of contexts at the max indel sequence length.}

\item{n.bases.mh.cap}{Specifies the max bases in microhomology to consider when counting repeat and microhomology
contexts. Counts of longer indels will simply be binned to the counts of contexts at the max indel sequence length.}

\item{...}{Other arguments that can be passed to variantsFromVcf()}

\item{description}{Will return a 1-column matrix containing the absolute indel signature
contributions (i.e. the number of mutations contributing to each mutational signature). The
signatures used are insertions/deletions within repeat regions (ins.rep, del.rep),
insertions/deletions with flanking microhomology (ins.mh, del.mh), and insertions/deletions
which don't fall under the previous 2 categories (ins.none, del.none). Each category is further
stratified by the length of the indel.}
}
\value{
A 1-column matrix
}
\description{
Extract indel signatures
}
