% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsSv.R
\name{extractSigsSv}
\alias{extractSigsSv}
\title{Extract structural variant signatures}
\usage{
extractSigsSv(vcf.file, output = "signatures", sample.name = NULL,
  sv.caller = "manta", sv.len.cutoffs = c(10^3, 10^4, 10^5, 10^6, 10^7,
  Inf), signature.profiles = SV_SIGNATURE_PROFILES, ...)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{output}{Output the absolute signature contributions (default, 'signatures'), or the SV type/length contexts
('contexts')}

\item{sample.name}{If a character is provided, the header for the output matrix will be named to this. If none is
provided, the basename of the vcf file will be used.}

\item{sv.len.cutoffs}{SV length cutoff intervals as a numeric vector.}

\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are the mutation
contexts and the columns are  the mutational signatures.}
}
\value{
A 1-column matrix containing: (if output = 'signatures') the absolute signature contributions (i.e. the number of mutations contributing to
each mutational signature) , or (if output = 'contexts') the mutation contexts.
}
\description{
Extract structural variant signatures
}
