% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitToSignatures.R
\name{fitToSignatures}
\alias{fitToSignatures}
\title{Linear least-squares (non-negative) fitting}
\usage{
fitToSignatures(signature.profiles, mut.context.counts, verbose = F)
}
\arguments{
\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are the mutation
contexts and the columns are  the mutational signatures.}

\item{verbose}{Show messages?}

\item{mut.contexts}{A vector of mutation contexts for fitting}
}
\value{
If vector is provided to mut.contexts, a vector returned containing the the absolute
contribution of each signature (i.e. the number of mutations contributing to each mutational
signature). If a matrix is provided, a matrix of absolute contributions is returned
}
\description{
Calculate the contribution of each mutational signature in a sample, given a vector of mutation contexts.
The eps() and lsqnonneg() functions have been extracted and adapted from the pracma R package.
}
