% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsIndel.R
\name{getContextsIndel}
\alias{getContextsIndel}
\title{Extract indel sequence, type and length}
\usage{
getContextsIndel(bed, ref.genome = DEFAULT_GENOME,
  get.other.indel.allele = F, verbose = F)
}
\arguments{
\item{bed}{A dataframe containing the columns: chrom, pos, ref, alt}

\item{ref.genome}{A character naming the BSgenome reference genome. Default is
"BSgenome.Hsapiens.UCSC.hg19". If another reference genome is indicated, it will also need to be
installed.}

\item{get.other.indel.allele}{Only applies when mode=='indel' For indels, some vcfs only report
the sequence of one allele (REF for deletions and ALT for insertions). If TRUE, the unreported
allele will be retrieved from the genome: a 5' base relative to the indel sequence. This base
will also be added to the indel sequence and the POS will be adjusted accordingly (POS=POS-1).}

\item{verbose}{Print progress messages?}
}
\value{
A dataframe in the same structure as a bed file
}
\description{
Extract indel sequence, type and length
}
