% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsSnv.R
\name{getContextsSnv}
\alias{getContextsSnv}
\title{Extract trinucleotide contexts}
\usage{
getContextsSnv(df, ref.genome = DEFAULT_GENOME, verbose = F)
}
\arguments{
\item{df}{A dataframe containing the columns: chrom, pos, ref, alt}

\item{ref.genome}{A BSgenome reference genome. Default is BSgenome.Hsapiens.UCSC.hg19. If another
reference genome is indicated, it will also need to be installed.}

\item{verbose}{Print progress messages?}
}
\value{
A dataframe in the same structure as a bed file with an extra column stating the context
of each variant
}
\description{
Extract trinucleotide contexts
}
