% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsSv.R
\name{getContextsSv}
\alias{getContextsSv}
\title{Determine SV type and length}
\usage{
getContextsSv(df, sv.caller = "gridss", return.raw = F, verbose = F)
}
\arguments{
\item{df}{A dataframe containing the columns: chrom, pos, ref, alt, filter, id, info}

\item{sv.caller}{Only applies when mode=='sv'. At this moment supports 'manta','manta_bnd,
'gridss', or 'pcawg'.
Currently there is no standard to how SVs are reported in vcfs. Therefore, output from different
callers will need to be parsed separately.}

\item{verbose}{Print progress messages?}
}
\value{
A dataframe in the same structure as a bed file with an extra column stating the context
of each variant
}
\description{
Determine SV type and length
}
