% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVcfFields.R
\name{getInfoValues}
\alias{getInfoValues}
\title{Get values from INFO field}
\usage{
getInfoValues(v, keys)
}
\arguments{
\item{v}{A character vector of the INFO field, with each item being a line of the INFO field}

\item{keys}{A character vector of the names of the INFO field values to retrieve}
}
\value{
A character vector (if 1 key was provided) or dataframe (if >1 key was provided)
containing the key names and corresponding values
}
\description{
Get values from INFO field
}
