% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIndelSignatureProfiles.R
\docType{data}
\name{INDEL_SIGNATURE_PROFILES}
\alias{INDEL_SIGNATURE_PROFILES}
\title{Indel signature profiles}
\format{
An object of class \code{matrix} with 83 rows and 18 columns.
}
\usage{
data(INDEL_SIGNATURE_PROFILES)
}
\description{
A matrix
rows: indel context
cols: indel signature
}
\details{
Source: https://cancer.sanger.ac.uk/cosmic/signatures/ID/index.tt
Each signature profile was downloaded manually, e.g. at https://cancer.sanger.ac.uk/sigs-assets-20/ID_vignettes/sigProfiler_ID_signatures_ID1.csv
}
\keyword{datasets}
