% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detHomology.R
\name{indelSeqFlanksStartEnd}
\alias{indelSeqFlanksStartEnd}
\title{Get the start/end positions for the left/right flanks of an indel}
\usage{
indelSeqFlanksStartEnd(
  chrom,
  pos,
  indel.len,
  indel.type,
  n.indel.lengths.l = 1,
  n.indel.lengths.r = 1
)
}
\arguments{
\item{chrom}{A character stating the chromosome}

\item{pos}{The REF position as an integer}

\item{indel.len}{The length of the indel as an integer}

\item{indel.type}{A character stating whether the indel is an insertion ('ins') or deletion ('del')}

\item{n.indel.lengths.l}{The length of the flanking sequence to return (measured in indel lengths)}

\item{n.indel.lengths.r}{See n.indel.lengths.l}
}
\value{
A vector of the left flank start/end position and right flank start/end position
}
\description{
Helper function for extractSigsIndel(), getting flank start/end positions first, then retrieving the sequences
using getSeq with a vectorized input improves speed significantly.
}
