% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitToSignatures.R
\name{lsqnonneg}
\alias{lsqnonneg}
\alias{lsqnonneg.default}
\alias{lsqnonneg.matrix}
\alias{lsqnonneg.data.frame}
\title{Non-negative least squares fitting}
\usage{
lsqnonneg(x, ...)

\method{lsqnonneg}{default}(mut.context.counts, signature.profiles)

\method{lsqnonneg}{matrix}(mut.context.counts, signature.profiles, verbose = F)

\method{lsqnonneg}{data.frame}(mut.context.counts, signature.profiles, verbose = F)
}
\arguments{
\item{mut.context.counts}{A numeric vector of mutation context counts}

\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are
the mutation contexts and the columns are the mutational signatures.}
}
\value{
A vector returned containing the the absolute contribution of each signature
}
\description{
Ripped least squares fitting function from pracma
}
