% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBasesMH.R
\name{nBasesMH}
\alias{nBasesMH}
\title{Calculate the number of bases in microhomology}
\usage{
nBasesMH(indel.seq, flank.seq, indel.len)
}
\arguments{
\item{indel.seq}{The indel sequence as a character}

\item{flank.seq}{The flanking sequence as a character}
}
\value{
An integer stating the number of bases in microhomology
}
\description{
Helper function for extractSigsIndel(). Scans a maximum of 1 indel length from 5' to 3' in the flanking sequence to
detect identical bases. Stops prematurely if a non-identical base is found.
}
