% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detHomology.R
\name{nCopiesAlongFlank}
\alias{nCopiesAlongFlank}
\title{Calculate the number of copies of the indel sequence are present in the flank sequence}
\usage{
nCopiesAlongFlank(indel.seq, flank.seq)
}
\arguments{
\item{indel.seq}{The indel sequence as a character}

\item{flank.seq}{The flanking sequence as a character}
}
\value{
An integer stating the number of copies found in the flanking sequence
}
\description{
Helper function for extractSigsIndel(). Scans along the flanking sequence for
copies of the indel sequence using a sliding window with a length equal to the length of
the indel sequence. The sliding window jumps in increments equal to the indel length.
Scanning stops prematurely when the indel sequence does not match the one in the sliding
window. The indel sequence itself counts as one copy.

Scanning needs only to be done from the 5' -> 3' direction as the reported POS of an indel
in a repeat region is always the first position of the repeat region.
}
