% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContexts.R
\name{plotContexts}
\alias{plotContexts}
\title{Plot contexts}
\usage{
plotContexts(
  x,
  mut.type = "auto",
  group = "group1",
  y.axis.var.scale = T,
  horizontal.group.labels = F,
  force.group.labels = F
)
}
\arguments{
\item{x}{A matrix or vector of context counts}

\item{mut.type}{A string specifying the mutation type Can be 'snv', 'dbs', 'indel'}

\item{group}{A vector indicating the which group each sample belongs to. If provided, a plot will
be produced for each group. If unspecified, the function will assume that all rows of `x` belong
to the same group. If NULL, all rows will be considered as separate samples}

\item{y.axis.var.scale}{If TRUE, variable y-axis scale for each sample group will be used}

\item{horizontal.group.labels}{If TRUE, horizontal text for the sample group labels will be used}

\item{force.group.labels}{If TRUE, sample group label will be shown even if there is only one
group}
}
\value{
A ggplot2 object
}
\description{
Plot contexts
}
