% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContrib.R
\name{plotContrib}
\alias{plotContrib}
\title{Plot context or signature contributions}
\usage{
plotContrib(
  x,
  mode = "contexts",
  mut.type = "auto",
  group = NULL,
  y.axis.var.scale = F
)
}
\arguments{
\item{x}{A matrix or vector of context or signature contributions}

\item{mode}{Can be 'signatures' or 'contexts'}

\item{mut.type}{A string specifying the mutation type Can be 'snv', 'dbs', 'indel'}

\item{group}{A vector indicating the which group each sample belongs to. If provided, a plot will
be produced for each group. If unspecified, the function will assume that all rows of `x` belong
to the same group}
}
\value{
A ggplot2 object
}
\description{
Plot context or signature contributions
}
