% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSigsAsDf.R
\name{readSigsAsDf}
\alias{readSigsAsDf}
\title{Read multiple files outputed my mutSigExtractor into a data frame}
\usage{
readSigsAsDf(dir = NULL, files = NULL, samples.as.rows = T, verbose = T, ...)
}
\arguments{
\item{dir}{A directory from which to read the signatures/contexts}

\item{files}{Alternatively, a list of file paths can also be specified}

\item{samples.as.rows}{Samples as rows and signatures/contexts as columns?}

\item{verbose}{Show file reading progress?}

\item{...}{Arguments that can be pass to list.files() and read.table()}
}
\value{
A data frame
}
\description{
Read multiple files outputed my mutSigExtractor into a data frame
}
\examples{
base_dir <- '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/Luan_projects/CHORD/Breast_Organoids/matrices/'
contexts <- list(
   snv = readSigsAsDf(paste0(base_dir,'/snv_contexts')),
   indel = readSigsAsDf(paste0(base_dir,'/indel')),
   sv = readSigsAsDf(paste0(base_dir,'/sv_contexts'))
)
}
