% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSbsSignatureProfilesV2.R
\docType{data}
\name{SBS_SIGNATURE_PROFILES_V2}
\alias{SBS_SIGNATURE_PROFILES_V2}
\title{SBS signature profiles (original 30)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 96 rows and 30 columns.
}
\usage{
data(SBS_SIGNATURE_PROFILES_V2)
}
\description{
A matrix
rows: 96-trinucleotide context
cols: 30 single base substitution (SBS) signatures
}
\details{
Source: https://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt
}
\keyword{datasets}
