% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSbsSignatureProfilesV3.R
\docType{data}
\name{SBS_SIGNATURE_PROFILES_V3}
\alias{SBS_SIGNATURE_PROFILES_V3}
\title{SBS signature profiles (new; v3, May 2019)}
\format{
An object of class \code{data.frame} with 96 rows and 60 columns.
}
\usage{
data(SBS_SIGNATURE_PROFILES_V3)
}
\description{
A matrix
rows: 96-trinucleotide context
cols: single base substitution (SBS) signatures
}
\details{
Source: https://cancer.sanger.ac.uk/sigs-assets-20/COSMIC_Mutational_Signatures_v3.1.xlsx
}
\keyword{datasets}
