% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSnvSignatureProfiles.R
\docType{data}
\name{SNV_SIGNATURE_PROFILES}
\alias{SNV_SIGNATURE_PROFILES}
\title{SNV signature profiles}
\format{An object of class \code{matrix} with 96 rows and 30 columns.}
\usage{
data(SNV_SIGNATURE_PROFILES)
}
\description{
A matrix
rows: 96-trinucleotide context
cols: 30 SNV signatures
}
\details{
Source: https://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt
}
\keyword{datasets}
