% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDfRegex.R
\name{splitDfRegex}
\alias{splitDfRegex}
\title{Split data frame or matrix using regex on colnames}
\usage{
splitDfRegex(df, patterns, return.remainder = T, names = NULL)
}
\arguments{
\item{df}{A data frame of matrix}

\item{patterns}{A vector/list of regular expressions. If arguments are named, or items in the
vector/list are named, the items in the output list will also be named accordingly}

\item{return.remainder}{Return columns which don't match the given regular expressions?}
}
\value{
A list of data frames or matrices
}
\description{
Split data frame or matrix using regex on colnames
}
\examples{
## Will return a list containing the corresponding names
splitDfRegex(df = contexts, patterns = list(snv = '\\\\[', indel = '[.]', sv = '[A-Z]{3}'))

## The same output list as above, but without names
splitDfRegex(df = contexts, patterns = list('\\\\[', '[.]', '[A-Z]{3}'))
}
