% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformContexts.R
\name{transformContexts}
\alias{transformContexts}
\title{Simplify and normalize contexts}
\usage{
transformContexts(contexts = NULL, snv = NULL, indel = NULL,
  sv = NULL, simplify.types = NULL, rel.types = NULL)
}
\arguments{
\item{contexts}{A list containing data frames of snv, indel, and/or sv contexts. Alternatively,
These can be specified with the arguments: snv, indel, or sv}

\item{snv}{See argument 'contexts'.}

\item{indel}{See argument 'contexts'.}

\item{sv}{See argument 'contexts'.}

\item{simplify.types}{Which types to flatten. Accepts: 'snv','indel','sv'}

\item{rel.types}{Which types to convert to relative contribution. Accepts: 'snv','indel','sv'}
}
\value{
A matrix or data frame
}
\description{
Exprimental; use at your own risk!
}
\examples{
base_dir <- '/Users/lnguyen/hpc/cog_bioinf/cuppen/project_data/Luan_projects/CHORD/Breast_Organoids/matrices/'
contexts <- list(
   snv = readSigsAsDf(paste0(base_dir,'/snv_contexts')),
   indel = readSigsAsDf(paste0(base_dir,'/indel')),
   sv = readSigsAsDf(paste0(base_dir,'/sv_contexts'))
)
transformContexts(contexts, simplify.types = c('snv','indel','sv'), rel.types = c('snv','indel','sv'))
}
