% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantsFromVcf.R
\name{variantsFromVcf}
\alias{variantsFromVcf}
\title{Extract relevant variant info for extracting SNV, indel, and SV signatures.}
\usage{
variantsFromVcf(vcf.file, mode = NULL, get.other.indel.allele = F,
  sv.caller = "manta", ref.genome = DEFAULT_GENOME,
  chrom.group = "all", vcf.filter = NULL, verbose = F)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{mode}{A character stating which type of signature is to be extracted: 'snv','indel', or 'sv'}

\item{get.other.indel.allele}{Only applies when mode=='indel' For indels, some vcfs only report
the sequence of one allele (REF for deletions and ALT for insertions). If TRUE, the unreported
allele will be retrieved from the genome: a 5' base relative to the indel sequence. This base
will also be added to the indel sequence and the POS will be adjusted accordingly (POS=POS-1).}

\item{sv.caller}{Only applies when mode=='sv'. At this moment supports 'manta' or 'gridss'.
Currently there is no standard to how SVs are reported in vcfs. Therefore, output from different
callers will need to be parsed separately.}

\item{ref.genome}{A character naming the BSgenome reference genome. Default is
"BSgenome.Hsapiens.UCSC.hg19". If another reference genome is indicated, it will also need to be
installed.}

\item{chrom.group}{chrom.group can be 'auto' for autosomes, 'sex' for sex chromosomes/allosomes,
'circular' for circular chromosomes. The default is 'all' which returns all the chromosomes.}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{verbose}{Print progress messages?}
}
\value{
A data frame containing the relevant variant info for extracting the indicated signature type
}
\description{
Extract relevant variant info for extracting SNV, indel, and SV signatures.
}
