% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosSim.R
\name{cosSim}
\alias{cosSim}
\alias{cosSim.default}
\alias{cosSim.matrix}
\alias{cosSim.data.frame}
\title{Cosine similarity}
\usage{
cosSim(x, ...)

\method{cosSim}{default}(x, y)

\method{cosSim}{matrix}(x, y)

\method{cosSim}{data.frame}(x, y)
}
\arguments{
\item{x}{A numeric vector or matrix}

\item{y}{A numeric vector or matrix}
}
\value{
A numeric vector or matrix of cosine similarities
}
\description{
Calculate cosine similarity between two vectors of the same length, or row-wise
cosine similarity of two matrices of the same dimensions
}
\examples{
set.seed(1)
x <- matrix(runif(200), nrow=20)
y <- matrix(runif(200), nrow=20)

## Works with matrices
cosSim(x, y)

## ...or vectors
cosSim(x[1,], y[1,])

}
