% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDbsSignatureProfiles.R
\docType{data}
\name{DBS_SIGNATURE_PROFILES}
\alias{DBS_SIGNATURE_PROFILES}
\title{DBS signature profiles}
\format{
An object of class \code{matrix} with 78 rows and 11 columns.
}
\usage{
data(DBS_SIGNATURE_PROFILES)
}
\description{
A matrix
rows: dinucleotide base substitution (DBS) context
cols: DBS signatures
}
\details{
Source: https://www.synapse.org/#!Synapse:syn12009743
}
\keyword{datasets}
