% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSigsSv.R
\name{extractSigsSv}
\alias{extractSigsSv}
\title{Extract structural variant signatures}
\usage{
extractSigsSv(
  vcf.file = NULL,
  df = NULL,
  output = "signatures",
  sample.name = NULL,
  sv.caller = "gridss",
  half.tra.counts = F,
  sv.len.cutoffs = if (output == "signatures") {     c(10^c(3:7), Inf) } else {    
    c(0, 10^c(3:7), Inf) },
  signature.profiles = SV_SIGNATURE_PROFILES,
  verbose = F,
  ...
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{df}{A dataframe with the columns: sv_type, sv_len. sv_type can be DEL, DUP, INV, TRA, BND.
Note that for TRA and BND, sv_len will be ignored. Alternative input option to vcf.file}

\item{output}{Output the absolute signature contributions (default, 'signatures'), or the SV
type/length contexts ('contexts')}

\item{sample.name}{If a character is provided, the header for the output matrix will be named to
this. If none is provided, the basename of the vcf file will be used.}

\item{sv.caller}{Can be 'manta', 'gridss', or 'pcawg'}

\item{half.tra.counts}{Divide translocation counts by 2?}

\item{sv.len.cutoffs}{SV length cutoff intervals as a numeric vector.}

\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are
the mutation contexts and the columns are  the mutational signatures.}

\item{verbose}{Print progress messages?}
}
\value{
A 1-column matrix containing the context counts or signature contributions
}
\description{
Will return a 1-column matrix containing: (if output = 'signatures') the absolute
signature contributions (i.e. the number of mutations contributing to each mutational signature),
or (if output = 'contexts') the mutation contexts.

To elaborate, the 6 SV signatures used are those described in this paper: https://media.nature.com/original/nature-assets/nature/journal/v534/n7605/extref/nature17676-s3.zip,
in Supplementary.Table.21.Signatures.v3.xlsx. These are derived from mutation contexts composed
of SV type/length.

Note that the probabilities of the clustered and non-clustered rearrangements in the signature
profile have been combined. In other words, whether the rearrangements were
clustered/non-clustered were not considered.
}
