% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIndelSignatureProfiles.R
\docType{data}
\name{INDEL_SIGNATURE_PROFILES}
\alias{INDEL_SIGNATURE_PROFILES}
\title{Indel signature profiles}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 83 rows and 15 columns.
}
\usage{
data(INDEL_SIGNATURE_PROFILES)
}
\description{
A matrix
rows: indel context
cols: indel signature
}
\details{
Source: https://cancer.sanger.ac.uk/sigs-assets-20/COSMIC_Mutational_Signatures_v3.1.xlsx
}
\keyword{datasets}
