% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitToSignatures.R
\name{lsqnonneg}
\alias{lsqnonneg}
\title{Non-negative least squares fitting}
\usage{
lsqnonneg(mut.context.counts, signature.profiles)
}
\arguments{
\item{mut.context.counts}{A numeric vector of mutation context counts}

\item{signature.profiles}{A matrix containing the mutational signature profiles, where rows are
the mutation contexts and the columns are the mutational signatures.}
}
\value{
A vector returned containing the the absolute contribution of each signature
}
\description{
Ripped least squares fitting function from pracma
}
