% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVcfFields.R
\name{readVcfFields}
\alias{readVcfFields}
\title{Read vcf into R as data frame}
\usage{
readVcfFields(vcf.file, fields = NULL)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{fields}{A character or integer vector indicating which columns to keep.}
}
\value{
A dataframe
}
\description{
Read vcf into R as data frame
}
\examples{
readVcfFields('/path/to/vcf', fields=c('CHROM','POS','REF','ALT'))
readVcfFields('/path/to/vcf', fields=c(1,2,4,5))
}
