% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSbsSignatureProfilesV2.R,
%   R/dataSbsSignatureProfilesV3.R
\docType{data}
\name{SBS_SIGNATURE_PROFILES_V2}
\alias{SBS_SIGNATURE_PROFILES_V2}
\title{SBS signature profiles (original 30)}
\format{An object of class \code{matrix} with 96 rows and 30 columns.}
\usage{
data(SBS_SIGNATURE_PROFILES_V2)

data(SBS_SIGNATURE_PROFILES_V2)
}
\description{
A matrix
rows: 96-trinucleotide context
cols: 30 single base substitution (SBS) signatures

A matrix
rows: 96-trinucleotide context
cols: single base substitution (SBS) signatures
}
\details{
Source: https://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt

Source: https://www.synapse.org/#!Synapse:syn12009743
}
\keyword{datasets}
