% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSmnvs.R
\name{subsetSmnvs}
\alias{subsetSmnvs}
\title{Subset for SNVs, DBSs, or indels}
\usage{
subsetSmnvs(df, type, chrom.name.style = "UCSC", verbose = F)
}
\arguments{
\item{df}{A dataframe containing the columns chrom, pos, ref, alt}

\item{type}{Can be 'snv', 'indel', or 'dbs'}

\item{chrom.name.style}{A string indicating the chromosome naming style. This can be for example
'UCSC' ('chrX') or 'NCBI' ('X'). See the documentation for `GenomeInfoDb::seqlevelsStyle()` for
more details.}

\item{verbose}{Show progress messages?}
}
\value{
The filtered dataframe, or an empty dataframe if no variants exist/remain
}
\description{
Standard variant filtering and subsetting for variant type used to clean the input
dataframe for extractSigs{Snv,Indel,Dbs}() functions
}
