% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantsFromVcf.R
\name{variantsFromVcf}
\alias{variantsFromVcf}
\title{Extract relevant variant info for extracting SNV, indel, and SV signatures.}
\usage{
variantsFromVcf(
  vcf.file,
  vcf.filter = NA,
  vcf.fields = c("CHROM", "POS", "REF", "ALT", "FILTER"),
  keep.chroms = NULL,
  chrom.name.style = "UCSC",
  merge.consecutive = F,
  verbose = F
)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{vcf.filter}{A character or character vector to specifying which variants to keep,
corresponding to the values in the vcf FILTER column}

\item{vcf.fields}{A character vector specifying the vcf columns to retrieve}

\item{keep.chroms}{A character vector specifying which chromosomes to keep (chromosome names
should be in the style of the vcf). To keep autosomal and sex chromosomes for example use:
keep.chroms=c(1:22,'X','Y')}

\item{chrom.name.style}{A string indicating the chromosome naming style. This can be for example
'UCSC' ('chrX') or 'NCBI' ('X'). See the documentation for `GenomeInfoDb::seqlevelsStyle()` for
more details.}

\item{merge.consecutive}{Some vcfs report MNVs as consecutive variants. For these vcfs, such rows
need to be merged into one row for proper function of downstream mutSigExtractor functions.}

\item{verbose}{Print progress messages?}
}
\value{
A data frame with each vcf field as a column
}
\description{
Extract relevant variant info for extracting SNV, indel, and SV signatures.
}
