% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantsFromVcf.R
\name{variantsFromVcf}
\alias{variantsFromVcf}
\title{Extract relevant variant info for extracting SNV, indel, and SV signatures.}
\usage{
variantsFromVcf(vcf.file, mode = NULL, sv.caller = "manta",
  ref.genome = DEFAULT_GENOME, chrom.group = "all",
  vcf.filter = NULL, verbose = F)
}
\arguments{
\item{vcf.file}{Path to the vcf file}

\item{mode}{A character stating which type of signature is to be extracted. SNV: 'snv', indel: 'indel', SV: 'sv'.
If unspecified, CHROM, POS, REF, and ALT will be returned}

\item{ref.genome}{A character naming the BSgenome reference genome. Default is "BSgenome.Hsapiens.UCSC.hg19". If another
reference genome is indicated, it will also need to be installed.}

\item{chrom.group}{chrom.group can be 'auto' for autosomes, 'sex' for sex chromosomes/allosomes, 'circular' for circular
chromosomes. The default is 'all' which returns all the chromosomes.}

\item{vcf.filter}{A character or character vector to specifying which variants to keep, corresponding to the values in
the vcf FILTER column}

\item{verbose}{Print progress messages?}
}
\value{
A data frame containing the relevant variant info for extracting the indicated signature type
}
\description{
Extract relevant variant info for extracting SNV, indel, and SV signatures.
}
